# Conflicts:
#	src/main/java/com/green/backend/application/entity/Application.java

+++ src/main/java/com/green/backend/expertreport/repository/ExpertReportRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExpertReportRepository extends JpaRepository<ExpertReport, Integer> {
+public interface ExpertReportRepository extends JpaRepository<ExpertReport, Long> {
 }
